requirements.txt
requests~=2.31.0
beautifulsoup4~=4.12.3
httpx~=0.27.0
pillow~=10.2.0
openai~=1.14.1
anthropic~=0.20.0
pywebcopy~=7.0.2
vertexai~=1.43.0
json2html~=1.3.0